requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
alembic==1.13.1
pyodbc==5.1.0
pandas==2.2.0
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.1.0
python-multipart==0.0.9
